renaming font_rtt cache to manager. absorbing related font functionality spread within canvas impl
