+++ CLAUDE.md
@@ -422,13 +422,13 @@ canvas.renderImguiTo(renderTargetId, canvas.defaultImguiFontId(), [&] { ... });
 - The OpenGL ImGui backend is FBO-agnostic; the Vulkan backend's per-context pipeline is built against `mRttRenderPass`, so it is render-pass-compatible with `beginRttPass`. **No changes to `imgui_impl_opengl3.*` / `imgui_impl_vulkan.*` are required.**
 - The platform backend (GLFW/SDL3) is skipped for secondary contexts — they run headless-style with `IO.DisplaySize` / `IO.DeltaTime` set manually. This means the feature works identically across GLFW+OpenGL, GLFW+Vulkan, SDL3+OpenGL, SDL3+Vulkan, and headless Vulkan.
 
-**Font handling — `ImguiRttFontCache` and `ImguiFontId`:**
+**Font handling — `ImguiFontManager` and `ImguiFontId`:**
 
-`ImguiRttFontCache` ([source/imgui_rtt_font_cache.h](source/imgui_rtt_font_cache.h), held by `ImguiRttManager`) owns each font as an entry in an `IndexedContainer<FontEntry>`, keyed by the public `ImguiFontId` ([include/imgui_font_id.h](include/imgui_font_id.h)). Atlas glyphs are owned by the shared `ImFontAtlas`; the cache stores non-owning observer pointers and patches them in place when the atlas rebuilds.
+`ImguiFontManager` ([source/imgui_font_manager.h](source/imgui_font_manager.h), held by `ImguiRttManager`) owns the entire ImGui-font lifecycle for a Canvas: the main HiDPI font (used by main-canvas UI), the secondary-context default font, every user-baked size, and the deferred bake/remove queue + atlas-rebuild flow. Each user-baked font is an entry in an `IndexedContainer<FontEntry>`, keyed by the public `ImguiFontId` ([include/imgui_font_id.h](include/imgui_font_id.h)). Atlas glyphs are owned by the shared `ImFontAtlas`; the manager stores non-owning observer pointers and `rebakeAll()` patches them in place across rebuilds.
 
-- At canvas construction, the cache is bound to the embedded Roboto TTF buffer (RAII via `ImguiRttManager` ctor). It bakes one font at the unscaled `imguiFontSize`, registers it as `io.FontDefault` for every secondary RTT context, and exposes its id via `Canvas::defaultImguiFontId()`. Result: ImGui text inside an RTT renders at its logical pixel height *in RTT pixels* — OS DPI scaling has no meaning in the game-canvas pixel grid, and the secondary-context default deliberately ignores it.
-- `Canvas::bakeImguiFont(float sizePx) -> ImguiFontId` ([include/canvas.h](include/canvas.h)) bakes additional sizes on demand. Repeat calls with the same `sizePx` return the same id (the cache dedups by size — ImGui itself does not dedupe `AddFontFromMemoryTTF` calls).
-- `Canvas::removeImguiFont(ImguiFontId)` schedules a full atlas rebuild at the start of the next frame (`ImFontAtlas::Clear()` + re-add main HiDPI font + `ImguiRttFontCache::rebakeAll()` + GPU font texture re-upload via `ActiveBackend::rebuildImguiFontTexture()`). The id passed to remove is invalidated; every other id survives the rebuild because the cache patches each entry's `ImFont*` in place — `Canvas::renderImguiTo(rtId, otherId, cb)` keeps working without intervention.
+- At canvas construction, `ImguiFontManager::initialize(contentScale)` adds the main HiDPI font at `imguiFontSize * contentScale * contentScale` for crisp DPI-aware glyphs on the main UI, then bakes the unscaled `imguiFontSize` and registers it as `io.FontDefault` for every secondary RTT context. The default-font's id is exposed via `Canvas::defaultImguiFontId()`. Result: ImGui text inside an RTT renders at its logical pixel height *in RTT pixels* — OS DPI scaling has no meaning in the game-canvas pixel grid, and the secondary-context default deliberately ignores it.
+- `Canvas::bakeImguiFont(float sizePx) -> ImguiFontId` ([include/canvas.h](include/canvas.h)) bakes additional sizes on demand. Repeat calls with the same `sizePx` return the same id (the manager dedups by size — ImGui itself does not dedupe `AddFontFromMemoryTTF` calls).
+- `Canvas::removeImguiFont(ImguiFontId)` schedules a full atlas rebuild at the start of the next frame: `ImFontAtlas::Clear()` + re-add main HiDPI font + `rebakeAll()` for surviving entries + secondary-context `io.FontDefault` refresh + GPU font texture re-upload via `ActiveBackend::rebuildImguiFontTexture()`. Orchestration lives on `ImguiRttManager::drainPendingFontOps(contentScale)` (called once per frame from `runOneFrame`); the cache-side parts (Clear + re-add + rebake) live on `ImguiFontManager::drainPendingOpsAndRebuildAtlas(contentScale)`. The id passed to remove is invalidated; every other id survives the rebuild because each entry's `ImFont*` is patched in place — `Canvas::renderImguiTo(rtId, otherId, cb)` keeps working without intervention.
 - `Canvas::pushImguiFont(ImguiFontId) / popImguiFont()` mid-callback override the panel's font without touching `ImFont`. `Canvas::isImguiFontReady(id)` guards against the one-frame deferred-bake window after `bakeImguiFont` returns; `Canvas::getImguiFontPtr(id) -> ImFont*` is the escape hatch for ImGui APIs that take an `ImFont*` directly (`ImGui::CalcTextSizeA` etc.).
 - The ID-stable design: only `removeImguiFont(id)` invalidates `id`. Atlas rebuilds (triggered by any other id's removal) leave every other id valid — the underlying pointer changes but `Canvas::renderImguiTo`, `pushImguiFont`, etc. resolve through the id automatically. See [imgui_font_removal_analysis.md](imgui_font_removal_analysis.md) for the rationale behind eager full rebuild vs alternatives.
