+++ source/imgui_rtt_manager.cpp
@@ -14,10 +14,11 @@ void ImGuiContextDeleter::operator()(ImGuiContext* ctx) const noexcept
 ImguiRttManager::ImguiRttManager(ActiveBackend& backend,
                                   RenderTargetContainer& renderTargets,
                                   const void* fontData,
-                                  std::size_t fontDataLen)
+                                  std::size_t fontDataLen,
+                                  float imguiFontSize)
     : mBackend(backend),
       mRenderTargets(renderTargets),
-      mFonts(fontData, fontDataLen)
+      mFonts(fontData, fontDataLen, imguiFontSize)
 {}
 
 void ImguiRttManager::enqueue(RenderTargetId renderTargetId, ImguiDrawCallback imguiDrawCallback)
@@ -147,6 +148,14 @@ void ImguiRttManager::refreshSecondaryContextDefaultFont()
         ImGui::GetIO().FontDefault = newDefault;
     }
     ImGui::SetCurrentContext(mainCtx);
+}
+
+void ImguiRttManager::drainPendingFontOps(float contentScale)
+{
+    if (!mFonts.hasPendingOps()) return;
+    mFonts.drainPendingOpsAndRebuildAtlas(contentScale);
+    refreshSecondaryContextDefaultFont();
+    mBackend.rebuildImguiFontTexture();
 }
 
 }
